+++ frontend/src/app/modules/shared/shared.module.ts
@@ -32,6 +32,7 @@ import { SeveritySelectComponent } from '@shared/components/severity-select/seve
 import { SeverityComponent } from '@shared/components/severity/severity.component';
 import { TextWithIconComponent } from '@shared/components/text-with-icon/text-with-icon.component';
 import { NotificationModalContentComponent } from '@shared/modules/notification/modal/content/notification-modal-content.component';
+import { FlattenObjectPipe } from '@shared/pipes/flatten-object.pipe';
 import { FormatPaginationSemanticDataModelToCamelCasePipe } from '@shared/pipes/format-pagination-semantic-data-model-to-camelcase.pipe';
 import { FormatPartSemanticDataModelToCamelCasePipe } from '@shared/pipes/format-part-semantic-data-model-to-camelcase.pipe';
 import { FormatPartlistSemanticDataModelToCamelCasePipe } from '@shared/pipes/format-partlist-semantic-data-model-to-camelcase.pipe';
@@ -86,6 +87,7 @@ import { RequestAlertComponent } from '@shared/components/request-notification/r
     FormatPaginationSemanticDataModelToCamelCasePipe,
     FormatPartSemanticDataModelToCamelCasePipe,
     FormatPartlistSemanticDataModelToCamelCasePipe,
+    FlattenObjectPipe,
     ViewContainerDirective,
     AvatarComponent,
     LanguageSelectorComponent,
@@ -153,6 +155,7 @@ import { RequestAlertComponent } from '@shared/components/request-notification/r
         NotificationUserComponent,
         InputComponent,
         FormatPaginationSemanticDataModelToCamelCasePipe,
+        FlattenObjectPipe,
         FormatPartSemanticDataModelToCamelCasePipe,
         FormatPartlistSemanticDataModelToCamelCasePipe,
     ],