+++ CHANGELOG.md
@@ -36,11 +36,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Updated owasp:dependency-check from 8.3.1 to 8.4.0
 - Updated commons-io from 2.11.0 to 2.13.0
 - Updated snakeyaml from 2.0 to 2.1
+- Split up Parts View in Frontend to parts asBUilt and asPlanned
 - Replaced own implementation of getCatalog, negotiateAgreement, and validatePolicy with irs-client-library implementation.
 - Updated irs-registry-client from 1.1.0-SNAPSHOT to 1.2.0-SNAPSHOT
 - Updated irs-helm from 6.4.1 to 6.5.0
 - Migrated groovy integration tests to SpringBootTests
 - OwnPageable now contains filter param
+
 
 ### Removed